Apply a branch-specific NPM tag to release builds that are not `latest`.
We have an issue where, if we have already cut the release for the next
minor version before cutting a hotfix release for an existing production
version, the corresponding npm packages would be erroneously tagged as
latest. This is because npm infers the `latest` tag if one is not
supplied.

Now, releases can receive a second tag that marks them as the latest in
a specific branch (e.g., `latest-0.84`).