adapter: Prevent EXPLAIN from being linearized (#24828)
If a peek is executed at a timestamp larger than the timestamp oracle's
read timestamp, then that peek must be linearized. Linearized here means
that the peek must wait for the timestamp oracle's read timestamp to
catch up to the timestamp of the read before returning a result to the
client.

Previously, Since `EXPLAIN` and `EXPLAIN TIMESTAMP` share the same code
path as peeks, they also were linearized. This is unnecessary because
`EXPLAIN` and `EXPLAIN TIMESTAMP` don't actually return data at some
timestamp, but rather explain how a query would be executed at some
timestamp. This is especially bad for `EXPLAIN TIMESTAMP`, where the
user may want to know if a query would block without blocking the
`EXPLAIN TIMESTAMP` query. This commit fixes the issue by not
linearizing `EXPLAIN` and `EXPLAIN TIMESTAMP` statements.