test: Fix test_leader_promotion test
The comments in the `test_leader_promotion` were not accurately
describing what the code was doing and half of the test was in an
unjoined thread which resulted in a lot of dead code. This commit
cleans up the test so the code matches the comments and it is actually
run.