catalog: Prevent persist catalog from early fence
All state that gets written to persist is tagged with the version of
the code that wrote that state. Persist has limited forward
compatibility in how many version in the future a reader can read.
Reading from persist updates state and the version that the state is
tagged with. As a consequence, reading from persist may unintentionally
fence out other readers and writers with a lower version. This is
especially dangerous with the catalog-debug tool, which may read from
persist at a higher version than is currently deployed in the database.

This commit adds machinery to the persist catalog so that readers
cannot accidentally fence other readers/writers, and writers cannot
increment the version in persist past what is allowable by the forward
compatibility guarantees.

Resolves #24695