persist-txn: switch to a new operator protocol for lazy
Instead of emitting the raw txns shard data, switch to a new protocol
based on mapping a "physical" upper to the "logical" upper that is known
to have exactly the same contents. Ideally, this would be exactly the
existing mz remap protocol (which has a nice formal definition), but we
can't quite get there yet, so that's left as a TODO.

The immediate motivation is to set up using TxnsRead in the first
operator (in a followup PR). This will greatly reduce our sensitivity to
things like the compaction of the txns shard because we'll go from one
(filtered) txns shard subscription per table input in a dataflow to one
(unfiltered) txns shard subscription total.

However, this does get us an immediate benefit of moving the
`unblock_reads` call from the second operator (which is run on each
worker) to the first operator (which is only run on one worker). This
hopefully helps with some excessive usage Aljoscha was seeing when
running local tests with many workers.