storage-controller: downgrade some asserts to soft_assert_or_log
assertions in the storage controller panic envd and prevent a
customer's environment from booting after e.g. version upgrades.
We have a customer whose environment is known to be in a state
that can trigger these asserts, but we do not want their environment
to be un-upgradeable. To resolve this, downgrade the asserts to
soft_assert_or_log so that we can both turn on their environment
and be notified if the assertion triggers (in this case via Sentry).