sql: add `PlanContext::ignore_if_exists_errors`
It could happen that we want to parse a DDL satement without erroring
out if the catalog item that this statement is meant to create already
exists (for example this happens when executing `EXPLAIN REPLAN`
statements). In such cases it is convenient to be able to ignore these
errors when configuring the parser, as opposed to re-implementing the
entire pipeline.