sql: remove Option wrapper around PostgresColumnDesc col_num
When introducing col_num, we needed to introduce it as an option
because existing sources did not have the field and it needed
to be backfilled. However, any subsources that did not have this
value backfilled by a migration would have errored as soon as it
produced data (we would have seen the schemas as incompatible).

This has been the case for 11 months now, so I suggest that this
is safe to remove because, for it to be an issue, customers would
have had a source stuck in an errored state for that duration of
time.