storage: remove flaky assertion
It turns out that `librdkafka` can send messages to the broker on its
own volition (probably on a timer from a background thread or something
like that), and those messages are counted by `in_flight_count`.

This means that we'll have to take `librdkafka`'s word for it and not
assert its claim ourselves. I have also manually audited its code and it
indeed flushes everything before committing.

Fixes #24721

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>