compute: fix non-aggressive readhold downgrades
The flag `enable_compute_aggressive_readhold_downgrades` was meant to
make it possible to disable aggressive readhold downgrading for MVs and
revert back to the old behavior where read holds are kept at the current
system time. As it turns out, the flag doesn't work as advertised
because adapter doesn't install read policies for MVs anymore, so when
you disable the flag, read frontiers of MVs and their inputs simply
become stuck.

To keep the changes at small as possible, this commit doesn't bring back
the read policy management for MVs in adapter. Instead it makes the
compute controller hold back the read frontier of sink dataflows to the
current time, as indicated by a `NowFn` passed in by adapter. This
doesn't exactly restore the previous behavior but enough to work around
make warming up of REFRESH MVs work again.

I expect that we can remove dependency on the `NowFn` again once we have
implemented #24966.