compute: fix non-aggressive readhold downgrades
The flag `enable_compute_aggressive_readhold_downgrades` was meant to
make it possible to disable aggressive readhold downgrading for MVs and
revert back to the old behavior where read holds are kept at the current
system time. As it turns out, the flag doesn't work as advertised
because adapter doesn't install read policies for MVs anymore, so when
you disable the flag, read frontiers of MVs and their inputs simply
become stuck.

This is fixed in this commit by making adapter also check the feature
flag and installing read policies if it is not set.

Note that adapter queries the flag's value from the compute controller,
rather than the system config. That's to ensure we don't run into
inconsistencies when the system config changes (i.e. someone runs `ALTER
SYSTEM SET`). Changing the system config for this flag only takes affect
after an environmentd restart, so we need to always base our decision on
the initial value.