compute: fix non-aggressive readhold downgrades
The flag `enable_compute_aggressive_readhold_downgrades` was meant to
make it possible to disable aggressive readhold downgrading for MVs and
revert back to the old behavior where read holds are kept at the current
system time. As it turns out, the flag doesn't work as advertised
because adapter doesn't install read policies for MVs anymore, so when
you disable the flag, read frontiers of MVs and their inputs simply
become stuck.

This is fixed in this commit by making adapter set read policies for MVs
again. These read policies are ignored when
`enable_compute_aggressive_readhold_downgrades` is set, but applied
otherwise, letting us revert to the old behavior.

This commit is mostly a revert of e6c647.