Revert compute aggressive read hold downgrades
This commit reverts the code that lets compute aggressively downgrade
read frontiers of sink dataflows. Doing so makes it impossible to warm
up REFRESH MVs, which we consider important for their usability. To make
warm up of REFRESH MVs work, we need adapter to continually keep the
read frontiers of them at a current time.