Fix implementation of RETAIN HISTORY for indexes (#24855)
We already supported `LOGICAL COMPACTION WINDOW` behind a feature flag,
but it has bit-rotted over the years and needed to be plumbed through to
a few more places.

Fixes: https://github.com/MaterializeInc/materialize/issues/24824

NOTE: we are still iterating on some questions like whether it's okay
that retaining history on one object implicitly retains history on its
dependencies. This PR just punts on that question.