adapter: use an empty compaction window for compute sinks
This commit makes adapter use an empty compaction window for compute
sink dataflows. It also changes the `lag_writes_by` read policy to
produce the write frontier when the given lag is zero.

The effect of these changes is that as-ofs of materialized views are
held back to a current time (the oracle read timestamp), which enables
warming up of REFRESH MVs by spawning their dataflows prior to the next
refresh time. At the same time the emtpy compaction window ensures that
MV read frontiers, and therefore the read frontiers of MV dependencies,
don't become stuck when an MV advances to the empty frontier.