Update when::Freshest to when::FreshestTableWrite
This commit updates the when::Freshest variant to
when::FreshestTableWrite. The issue with when::Freshest is that when
reading from a constant materialized view, we will choose
Timestamp::MAX as our timestamp. When we try and linearize this
timestamp, we end up waiting until the end of time, which takes awhile
to arrive.

The goal of when::Freshest was to avoid invalid retractions to tables
during read then write plans, by making sure we read the most recent
table writes. when::FreshestTableWrite advances to the timestamp
oracle's read timestamp, which is guaranteed to have all table writes,
but does not advance to the upper, which may be Timestamp::MAX for
constant views.