storage: determine output index by # of exports
The mechanism we use to demux multiple streams from a single
ingestion is referred to internally as its output index.

Previously the output index of a stream was determined in planning,
and, for multi-output ingestions, was based on the index of the
upstream table in the vec in which we kept its details.

This had the unfortunate side effect that it prevented us from
demuxing a single upstream table into multiple output indices.

This commit changes that relationship to instead determine the
output index of a collection based on the number of total
source exports.

To do this while still permitting steady references from a subsource
to an upstream table, we introduce a per-source type configuration
that correlates a subsource to an upstream table.

This has some long-term consequences in that it disallows us from
adding multiple subsources with the same upstream ID, but that
tradeoff seems reasonable.