environmentd: improve test connection polling
Previously a race condition between a client and server drop could cause the
connection half of the client to get `Some(Err(_))` messages, and then choose to
panic on those. However errors can happen here when the server side goes away,
and it's not necessarily panic-worthy (and perhaps should never panic). If the
server's ListenHandle dropped before the connection noticed that the client
dropped, this would cause panics.

This was happening for me locally. I'm confused why it is not happening in CI.