adapter: Fix linearization of read-then-write (#24862)
If a read is executed at a timestamps larger than the timestamp oracle's
read timestamp, then that read must be linearized. Linearized here means
that the read must wait for the timestamp oracle's read timestamp to
catch up to the timestamp of the read before returning a result to the
client.

Previously, read-then-write queries were mistakenly setting the current
timestamp oracle read timestamp to None. This caused the linearization
process to immediately return before properly linearizing the read. This
commit fixes the issue by accurately setting the timestamp oracle read
timestamp.