persist: remove some unnecessary intermediate Vecs
Instead of collecting an iterator into a `Vec` and then handing it to
`Antichain::from`, we can construct the Antichain directly from the
iterator. I did a quick audit and none of these seem to be on a
particularly hot path, but this is so easy that may as well.

h/t Moritz for the idea