storage: convert output_index to input_index
For sources with subsources, we determined the output index of the
source (i.e. the channel over which we demux the source output to
persist) based on the table's position in the upstream collcection,
e.g. the PG publication.

This prevents us from having a single table's values written to
multiple persist shards.

To prepare source to be able to demuz upstream tables into multiple
persist shards:

- Convert the output index into an input index, i.e. use the same
  information to correlate a subsource to an upstream table.
- Determine the output index using the stable enumeration of a
  BTreeMap that contains all of the outputs of the source. This
  provides a deterministic algorithm that can be used in the
  ingestion dataflows as well as rendering.

With this scheme, we will be able to have an m:n relationship
between upstream tables and persist shards.