compute-client: maintain warmup capabilities
This commit introduces warmup capabilities held by the controller for
compute collections. These capabilities hold back their collections'
read frontiers, and therefore the as-ofs of new dataflows installed for
collections, to a time that is immediately available in all inputs.

This is done to enable warming up dataflows, particularly REFRESH
materialized views that have their next refresh time in the future. By
holding back the as-ofs of dataflows we ensure that they can begin
hydration immediately and don't have to wait for a future time (like the
next refresh time) to do so.

Warmup capabilities are updated asynchonously as part of the periodic
controller maintenance. We can't update them synchronously because the
compute controller is not informed about updates to the write frontiers
of storage collections, which can cause updates to warmup capabilities.