test: make test_old_storage_usage_records_are_reaped_on_restart more lenient
This problem only became apparent when using the new Postgres-backed
timestamp oracle, which a previous commit makes the default.

Before, this was only working because on bootstrap we would advance the
boot_ts to the latest catalog oracle timestamp, which was 5000, and not
in fact 1100 as the test was assuming. The storage usage records that
were then pruned had (for reasons) a timestamp of 12.

When using the Postgres timestamp oracle, the boot_ts will not get
advanced to 5000. With this, the boot_ts stays at 1100 and so the
records with ts 12 will not be below the cutoff for pruning and will not
be pruned.

We now advance the "now" timestamp a bit further, to make sure that all
expected records are in fact pruned.