fivetran-destination: Support reporting `primary_key` from a `DescribeTable` request (#25031)
When we get a `CreateTableRequest` from Fivetran they include a list of
primary keys. Later when they run a `DescribeTableRequest` they expect
this information to roundtrip, i.e. if they made column "foo" a primary
key `DescribeTable` should indicate it as one. If `DescribeTable` does
not match their expectations they run an `ALTER TABLE ...` request to
update the table.

Materialize does not support primary keys, so to prevent Fivetran from
running an `ALTER TABLE ...` request (which we also don't yet support),
we store a magic string as a SQL COMMENT on the column to indicate
whether or not it's a primary key.

This is a HACK and is not something we want to ship with, instead we
should support `PRIMARY KEY ... NOT ENFORCED` and include whether or not
something is a primary key in the `mz_columns` table or something
similar. But in the interest of unblocking testing, this is the fastest
thing we could come up with.

### Motivation

Reported as a blocker by Fivetran:
https://materializeinc.slack.com/archives/C060KAR4802/p1707249394772389?thread_ts=1707202803.646169&cid=C060KAR4802

### Checklist

- [ ] This PR has adequate test coverage / QA involvement has been duly
considered.
- [ ] This PR has an associated up-to-date [design
doc](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/design/README.md),
is a design doc
([template](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/design/00000000_template.md)),
or is sufficiently small to not require a design.
  <!-- Reference the design in the description. -->
- [ ] If this PR evolves [an existing `$T ⇔ Proto$T`
mapping](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/command-and-response-binary-encoding.md)
(possibly in a backwards-incompatible way), then it is tagged with a
`T-proto` label.
- [ ] If this PR will require changes to cloud orchestration or tests,
there is a companion cloud PR to account for those changes that is
tagged with the release-blocker label
([example](https://github.com/MaterializeInc/cloud/pull/5021)).
<!-- Ask in #team-cloud on Slack if you need help preparing the cloud
PR. -->
- [x] This PR includes the following [user-facing behavior
changes](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/guide-changes.md#what-changes-require-a-release-note):
  - N/a