stash: Speed up large batch test (#25055)
test_stash_batch_large_number_updates was added in
dad4ed8c7124f5a5dd2e47f3597c7f98004407a0 to test the max number of args
in a CRDB update statement. Previously, we were testing a number of
updates that was order of magnitudes larger than the max number of args
in a CRDB update statement. That caused the test to be slow and
sometimes hang forever. This commit updates the test to lower the number
of updates while still ensuring that we end up with more than the max
number of args.

Works towards resolving #25003