adapter: Allow canceling linearized reads (#24826)
If a read is executed at a timestamps larger than the timestamp oracle's
read timestamp, then that read must be linearized. Linearized here means
that the read must wait for the timestamp oracle's read timestamp to
catch up to the timestamp of the read before returning a result to the
client. This situation is very rare and should only happen when the
since of some object goes past the oracle's read timestamp, which should
only happen due to clock skew.

Previously, while a read was being linearized (i.e. waiting), it was
impossible to cancel to read query. This commit fixes that issue by
allowing read queries to be canceled while they're being linearized.

The bulk of the work was storing the pending read transactions in a
field of the Coordinator instead of passing them around in channels and
tasks.