ore: add a clone of bytes::Bytes but backed by lgalloc
To set up moving allocations from our s3 library into lgalloc, we
introduce a copy of `bytes::Bytes` backed by lgalloc but with the same
API. Crucially, this also implements `bytes::Buf`. We could expand the
subset of the `bytes::Bytes` API that this implements, but start with
the bare minimum for doing something useful.

If we end up needing to fork our s3 client to push usage of this
earlier, we'll likely need to pull this out into a separate crate, but
for now start it in mz_ore.

This is intentionally not used anywhere yet. It's pulled into a separate
PR to keep code reviewer responsibilities less muddled.