Track the overall lower/upper in validations
Previously we were tracking the upper of the batch with the largest
lower, which is not obviously the same thing...

Also, the old code contained some PartialOrder::less_equal calls with
the result discarded. It's unclear to me what these were intended to
express; I just introduced an assert on the overall since.