adapter: general framework for off thread coordinator stages and spans
Make a new requirement that all stages but the last must return a
JoinHandle<Stage> or an ExecuteResponse. This allows us to remove the loop
(since it will never trigger) currently in use in all of the sequence stage fns,
and instead *only* have the final response or the next stage that can be sent as
a message.

Because of this, we can now omit a lot of passing of the vars down through the
various inner stage fns, and instead concentrate those in a common handling
function.

Additionally, plumb through spans and easy tracing for nicely formatted trees.

Start with CREATE VIEW for now as the smallest possible demonstration.