storage/mysql: leaderless snapshot
This PR removes the coordination of the snapshot phase and makes each
worker independently lock and snapshot its assigned tables. This has the
effect of simplifyin the code and potentially being more gentle to the
database by requiring locks of fewer tables for less amount of time at
any given point.

I also added a whole bunch of additional documentation on the mechanics
of how the snapshot phase works.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>