sql,adapter: `CREATE MATERIALIZED VIEW AS OF` syntax
This commit adds `AS OF` syntax to the `CREATE MATERIALIZED VIEW`
statement. In contrast to the existing syntax for `SELECT` and
`SUBSCRIBE` statements, this syntax is bare-bones and internal-only. It
is only meant to let us store the initial as-ofs of materialized views
in the catalog and consequently doesn't affect the timestamp selection
of created materialized views.