adapter: use initial MV as-of during bootstrap
This commit fixes a bootstrap as-of selection bug for materialized
views: If environmentd crashed after having committed a new MV to the
catalog but before having created the storage collection, bootstrapping
would initialize the storage collection with a `since` of `[0]`, making
the MV readable for times at which it didn't have valid contents.

The fix consists of writing the as-of selected during MV creation into
the `create_sql` that gets stored in the catalog. Then, during
bootstrap, the coordinator can pass that as-of value when creating the
storage collection, ensuring that it always gets a `since` of at least
the previously selected as-of.