catalog: Clean up boot_ts usage in open
This commit renames the `boot_ts` parameter of the open family of
method in `OpenableDurableCatalogState` to `initial_ts`. `initial_ts`
is more accurate because the timestamp is only used to initialize the
catalog if it's a new environment. Additionally, `boot_ts` is used in
other places and is linearized against the timestamp oracle. In the
open family of methods, `boot_ts` wasn't linearized, so the name may
have mistakenly led people to people that this `boot_ts` was also
linearized.

Additionally, `initial_ts` is removed as a parameter from
`open_readonly`, because `open_readonly` cannot initialize a new
environment, so it doesn't make sense to require a `initial_ts`.