test: Change default catalog to persist
There are some tests that cannot use the shadow catalog for various
reasons. Previously, those tests defaulted to using the stash catalog,
because that is the catalog used in production. We are actively working
on switching the stash catalog to the persist catalog in production,
and we're not changing the stash catalog at all. Therefore, we'd prefer
to test the persist catalog over the stash catalog.

Works towards resolving #24218