persist: use lgalloc to back data fetched from s3
This will reduce rehydration memory spikes by moving the data fetched
from s3 into file-backed mapped allocations.

Ideally, we'd read directly from the network into these allocs, but that
requires maintaining a fork of our s3 client library and/or hyper. We
appear to get a meaningful improvement by moving the data in at the very
first place it touches mz code, and that's a much easier place to start.
It would probably be more lgalloc-friendly to copy in a full s3 blob at
the end of the fetch, but initial testing on a staging env indicates
this is measurably worse than moving in each s3 multipart as it comes
in.

On its own, this seems to be sufficient to make parquet allocations the
limiting factor in rehydration spikes in practice. However, we may want
to supplement it with a semaphore around s3 fetching to get a hard upper
bound on how much data may have arrived from s3 but not yet been moved
into lgalloc. (This semaphore has the deficit of being similar in spirit
to the persist_source "physical backpressure" that we're trying to move
away from, but it has the benefit of working cross-persist_source, so it
may be worth doing in the short-term, anyway.)

Touches #23332