catalog-debug: Remove hard coded replica sizes (#25133)
Previously, the catalog-debug tool would hard-code replica sizes of `1`
in multiple places for the upgrade checker. This isn't always valid
because the size `1` is not valid for all environments (really for any
prod environments). This commit removes these hard coded values and
replaces them with more sensible values.

`default_cluster_replica_size` is only used for initializing new
environments. Specifically it is used for the size of the `quickstart`
cluster. The upgrade checker should never be used in non-existent
environments. So this commit replaces the size with the string "DEFAULT
CLUSTER REPLICA SIZE IS ONLY USED FOR NEW ENVIRONMENTS", so that we'll
get a meaningful error message if the value is ever used in the upgrade
checker.

`builtin_cluster_replica_size` is used as the size for new builtin
replicas, which is a valid scenario for an upgrade. This commit replaces
the size with some random valid size, so it doesn't fail. The actual
size doesn't matter because it's never persisted.