doc/user: add shortcodes for in-progress rollouts
We've had a lot of trouble lately with folks either:

  1. Not writing documentation for their changes.
  2. Writing documentation for their changes, but merging it in
     the PR that implements that change, before the change has rolled
     out to production.

Attempt to solve both of those problems by introducing new rollout
shortcodes. These shortcodes add a user-visible warning that indicates a
feature may not yet be available. The warning automatically disappears
when the version containing the feature is marked as released.

This will "bless" the approach of updating docs for a change in the same
PR that implements the change. This hopefully encourages developers to
write more documentation, without confusing end users with documentation
for features that do not yet exist.