optimize: add `reoptimize_imported_views` feature flag
Forcing `DataflowBuilder::reoptimize_imported_views` to reoptimize all
catalog-backed views imported in the `DataflowDescription` in every
possible `EXPLAIN` run is a bit of an overkill.

Instead, we should enable work in this method call only when one of the
overrides of `OptimizerConfig` (`CREATE CLUSTER ... FEATURES(...)` or
`EXPLAIN PLAN WITH(...)`) explicitly sets the value of a feature flag
that influences an optimizer stage that runs before we produce the
locally optimized `MirRelationExpr`.

To do this, we introduce a dedicated optimizer feature flag for that.
Engineers should enable this flag in conjunction with any other flags
that modify the view optimization pipeline in order to see differences
in catalog items that depend on the view.