catalog-debug: Add ignore cmnd line args to dump (#25136)
This commit adds the --ignore and --ignore-large-collections command
line arguments to the dump command in the catalog debug tool. Often
times when debugging, the user only cares about a handful of
collections, and the other collections can be annoying noise. The
--ignore flag allows the user to specify collections to ignore. The
--ignore-large-collections flag is a convenience flag to ignore both
audit_log and storage_usage, which are both extremely large and usually
not helpful.