storage: report an error when observing a legacy error
A few months ago we accidentally shipped a protobuf incompatible change
which was persisted in various collection shards. In order to handle
this we shipped a version of MZ[1] that automatically detects the
incompatible protobufs and heals the shards by retracting the
incompatible representation and inserting the correct one. The idea was
that eventually shard compaction would lead to the incompatible
protobufs getting deleted. As part of that work we also added metrics
tha tracked the occurances of legacy errors with the expection of that
metric dropping to zero after enough time had passed.

The metrics for the legacy representation of errors have indeed dropped
to zero for all environments in production. While this could be
considered enough evidence to finally remove the healing code I would
like to first get a release out that raises a sentry error if a legacy
representation is observed. If our metrics are accurate then this error
should not fire and in the release after that one we'll be free to
remove all this code.

[1] https://github.com/MaterializeInc/materialize/pull/21887

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>