Implement a chunked stack
Before this change, we used `TimelyStack`s to store region-allocated data.
TimelyStack only stores referenced objects in regions, but not the
immediate data itself. With this change, we region-allocate the `local`
portion of a stack, too, which allows us to use `lgalloc` for that memory.
The benefit is that more memory is eligible to be spilled to disk.

Signed-off-by: Moritz Hoffmann <mh@materialize.com>