persist: use lgalloc to back data decoded from arrow/parquet
This will reduce rehydration memory spikes by moving the data decoded
from arrow/parquet into file-backed mapped allocations.

Ideally, we'd read decode into these allocs, but that
requires maintaining a fork of our arrow client library. We
appear to get a meaningful improvement by moving the data in at the very
first place it touches mz code, and that's a much easier place to start.

Mechanically, this is accomplished by replacing the arrow2-provided
`*Buffer` structs (morally an `Arc<Vec<T>>`) with lgalloc-backed
`Regions`. On the decode path, we attempt to use `Region::Mmap` (only on
cc clusters). On the encode path, we assume the data will shortly be
encoded and don't bother, instead just using `Region::Heap`. This diff
is pretty small, most of the PR is generalizing `MetricsRegion` to any
`T: Copy` as well as plumbing the configs and metrics.

Along with #25097, this is sufficient to allow the TPCH (sf=1) queries
as indexes to rehydrate on a 3xsmall (4 GiB, modified with bonus disk).
That represents ~40 GiB of input data fetched from s3.

Touches #23332