sql: support parameters in LIMIT expressions
When the `enable_expressions_in_limit_syntax` feature flag is enabled,
this commit teaches Materialize to allow parameters in LIMIT
expressions.

The trick is to allow a `RowSetFinishing` to be generic over the type of
the limit, so that the planner can hold the finishing's limit as an
`HirScalarExpr` long enough to call `bind_parameters` on it.