test/cluster: unflake test-compute-controller-metrics
The compute controller now only refreshed its metrics once per second,
which makes it more likely that test-compute-controller-metrics, which
runs some SQL and then checks for that to be reflected in the metrics,
gets to see outdated values.

The only way we can solve this is by sleeping before querying the
metrics. Given the 1s refresh interval, sleeping for 2s should be
enough.