sql: Fix interval rounding behavior (#25202)
This commits fixes multiple bugs related to interval rounding. The
specific Interval rounding that is affecting is the rounding that
happens when parsing an Interval, for example
`INTERVAL '...' SECOND(4)'`, and the rounding that happens when calling
the `date_trunc` method with the seconds unit. The fixes are:

1. Previously, rounding of negative intervals was not handled at all.
This commit adds rounding for negative intervals.
2. Previously, rounding an interval could cause overflow/underflow. This
commit returns an error in those scenarios.
3. Previously, when `date_trunc` was called with the seconds unit,
microseconds were rounded to the nearest second. This commit updates the
logic so that microseconds are truncated which is the correct behavior
that matches PostgreSQL.

Fixes #25182