test: Fix flaky storage usage reap test (#25235)
This commit fixes the
test_old_storage_usage_records_are_reaped_on_restart test so that it's
no longer flakey. Previously, the test would query the timestamp from
the mz_storage_usage_by_shard table with the following query, `SELECT
EXTRACT(EPOCH FROM MAX(collection_timestamp))::integer FROM
mz_internal.mz_storage_usage_by_shard;`. The issue is that
`EXTRACT(EPOCH FROM <timestamp>)` returns a float8 that represents
seconds since the epoch, where the fractional component is sub-second.
When we cast to an integer, we were truncating the sub-second component.
Additionally, we were then treating the truncated value as a
milliseconds measurement when it was in fact a seconds measurement. This
commit fixes the issue, by multiplying the seconds by 1000 (the number
of milliseconds per second) before casting to an integer.

Fixes #25193