test: Simplify catalog snapshot tests
Multiple catalog tests use the insta crate to compare a snapshot of
the catalog contents against some snapshot file. The user_version is
a value in the catalog that frequently changes between versions. It
can be annoying to go through each snapshot file and change the
user_version to the correct value. This commit simplifies the process
by remove the user_version from the snapshot and comparing its value
in Rust against the expected value.