catalog: Fix persist data migration panics (#25277)
This commits fixes a panic that can happen in persist data migrations.
Before running migrations the catalog attempts to read the epoch and
config collection by calling `try_from` on the raw data to convert it
into structured data. This should fail on un-migrated collections by
returning an `Err`, not by panicking. However, the `try_from`
implementation was calling a `to_serde` method that panics when it
fails. This commit fixes the panic by adding a `try_to_serde` method
that returns an `Err` which can be propagated to `try_from`.