sql: Make mz_schema_oid more robust
This commit makes mz_schema_oid more robust by only searching for
schemas in the current database of the session when a database is not
explicitly specified. Additionally, this commit allows callers to
specify the database of a schema when calling mz_schema_oid. As a
result other functions that call mz_schema_oid, such as
has_schema_privilege, also are more robust.

Fixes #21490