adapter: add a SessionMeta struct for off-thread work
An upcoming refactor will move the ctx ownership during peeks out of the
individual stages, making it not possible to share the session across tasks. The
optimizer was designed to work across tasks, but we were breaking that
abstraction boundary by passing sessions by reference instead of some object
fully owned by the task.

Teach sessions how to generate a snapshot of their needed data for `fn
eval_unmaterializable_func` called a `SessionMeta`, a reference-less struct
containing mostly cheap data copies. It does have one exception to thas, which
is that `Var`s must allocate a new btree and value strings because `Var` is not
`Clone` (and is hard to make so), so isn't easily shareable. Overall we believe
this var snapshotting, even though it is done on every request, is still cheaper
than any other option we considered.

Abstract this new struct and the old Session with a common trait and use that in
places where it's supported. Due to vars being tricky some things do require a
SessionMeta, but we hope an upcoming `Var` refactor will allow this restriction
to be lifted.

Our goal is to make the coord thread is un-contested as possible, and thus move
work into off-thread tasks. If we did not make the current change, we would have
to do something like split up the optimize stage into two stages, and in between
them walk the expr and apply eval_unmaterializable_func to it on the main coord
thread (since it owns the Session). This seems like it'd be at least as much CPU
time as it takes to allocate a new btree and some Strings.

Other options we considered and rejected were:
- share the session with an `Arc` (bad because during ctx.retire and other
  operations we'd have to wait an unknown amount of time to be the only strong
  reference holder and confuses ownership responsibilities)
- use some immutable map data structure for session vars that has cheap
  snapshotting/cloning (not possible because values need to be `Clone`, and
  `Var` is not)
- change the sequence_staged API to pass around ctx by owner instead of
  reference (bad because its current API allows for regular `?` returns, thus
  not needing to wrap errors in `return_if_err!`, and has a single owner of ctx
  with really clear responsibilities about which component needs to retire it)