storage/source/mysql: fix snapshot transaction
The source snapshot code was inadvertently dropping the transaction it
intended to use on the floor, leading to the correctness issue observed
in #25124.

Specifically, the SELECT query to load the table's snapshot was running
outside of any transaction. This allowed it to observe the handful of
transactions that committed *after* the snapshot's consistent point was
constructed but before the SELECT query began executing. The replication
stream (correctly) did not rewind this data, resulting in the data from
that handful of transactions getting ingested into Materialize twice.

As far as I can tell, the code's approach to locking is (now) sound, so
we don't need any of the changes to the locking scheme proposed in #25300.
In particular, FLUSH TABLES only seems relevant if you need the data to
be flushed to disk--e.g., so that you can reach into MySQL's data
directory and tar up the files for export to another machine. Since we
export the data using a SELECT statement, it seems to me that using LOCK
should be sufficient [0].

The test enclosed in this PR is based on @rjobanp's test in #25300, but
modified to insert data in parallel with creating the source in
Materialize. The original test only failed sporadically on current main,
but the modified test failed reliably--i.e., in 100% of the 4-5 times I
ran it. The modified test also reliably passes with the fix in this
commit.

Closes #25300.
Fixes #25124.

[0]: Confirming that statement in the MySQL documentation is left as an
exercise to the reader.

Co-authored-by: Roshan Jobanputra <roshan@materialize.com>