compute-client: avoid regressing global write frontier
This commit fixes a bug in the compute controller where global write
frontiers could regress by being overwritten with the write frontiers of
lagging replicas.

The `maybe_update_global_write_frontiers` method did contain a check to
ensure that frontier regressions are ignored. However, this check was
applied only _after_ `CollectionState::write_frontier` was already
unconditionally updated to the new frontier, making it quite useless.

The fix is of course to check for regressions before updating the
collection state.