test: Fix legacy upgrade tests catalog impl (#25324)
This legacy upgrade test will skip versions when testing certain upgrade
paths. The persist catalog will panic if a version is skipped when
upgrading. Even when using the stash catalog implementation we open the
persist catalog during boot, which will cause the panic. To avoid this,
this commit uses the emergency-stash catalog implementation, which
doesn't even attempt to open the persist catalog.

Fixes #25321