compute-client: refactor `update_read_capabilities`
This method had accumulated some cruft other time and made some
assumptions that are outdated, such as:
 * `GlobalId`s have a meaningful order
 * Received `updates` can refer to storage collections

This mades the code difficult to reason about. Given how essential the
method is for the correct operation of the compute layer, it's worth
refactoring it for readability, which is what this commit does.

Apart from refactoring the internals, this also changes the arguments to
take an owned `BTreeMap`. I assume it was previously a reference because
we hoped to reuse the allocation when it was still a `HashMap` or a
`Vec`. Not much point in attempting the same with a `BTreeMap`. And none
of the callers wants to do so anyway.